Allow instance name to be set explicitly
and ensure it does not contain dots. The service name may include all sorts of characters except for the separator (the dot), and it may include spaces. If one wants to expose a service on a .local subdomain (for example - for Rails testing) - deriving the service name from the hostname _may_ work, but it is easier to make it settable explicitly. Devices like printers do that as well.

Another option would be to replace all dots in the service name before the service type with, say, dashes - that would be more automatic, but it is neat that the service name can be "art-directed".

Remove the .local component from the passed hostname, because it seems redundant - we only advertise the fully-qualified hostname, to which we append .local. anyway.